+++ apps/backoffice-v2/src/lib/blocks/hooks/useKybRegistryInfoBlock/useKybRegistryInfoBlock.tsx
@@ -1,9 +1,59 @@
-import { useMemo } from 'react';
+import { useCallback, useMemo } from 'react';
+
+import { WarningFilledSvg } from '@/common/components/atoms/icons';
 import { createBlocksTyped } from '@/lib/blocks/create-blocks-typed/create-blocks-typed';
 
 export const useKybRegistryInfoBlock = ({ pluginsOutput, workflow }) => {
+  const getCell = useCallback(() => {
+    if (Object.keys(pluginsOutput?.businessInformation?.data?.[0] ?? {}).length) {
+      return {
+        id: 'nested-details',
+        type: 'details',
+        hideSeparator: true,
+        value: {
+          data: Object.entries(pluginsOutput?.businessInformation?.data?.[0])?.map(
+            ([title, value]) => ({
+              title,
+              value,
+            }),
+          ),
+        },
+        workflowId: workflow?.id,
+        documents: workflow?.context?.documents,
+      } satisfies Extract<
+        Parameters<ReturnType<typeof createBlocksTyped>['addCell']>[0],
+        {
+          type: 'details';
+        }
+      >;
+    }
+
+    if (pluginsOutput?.businessInformation?.message) {
+      return {
+        type: 'paragraph',
+        value: (
+          <span className="flex text-sm text-black/60">
+            <WarningFilledSvg
+              className={'mr-[8px] mt-px text-black/20'}
+              width={'20'}
+              height={'20'}
+            />
+            <span>{pluginsOutput?.businessInformation?.message}</span>
+          </span>
+        ),
+      } satisfies Extract<
+        Parameters<ReturnType<typeof createBlocksTyped>['addCell']>[0],
+        {
+          type: 'paragraph';
+        }
+      >;
+    }
+  }, [pluginsOutput, workflow]);
+
   return useMemo(() => {
-    if (Object.keys(pluginsOutput?.businessInformation?.data?.[0] ?? {}).length === 0) {
+    const cell = getCell();
+
+    if (!cell) {
       return [];
     }
 
@@ -14,38 +64,22 @@ export const useKybRegistryInfoBlock = ({ pluginsOutput, workflow }) => {
         value: createBlocksTyped()
           .addBlock()
           .addCell({
-            type: 'container',
-            value: createBlocksTyped()
-              .addBlock()
-              .addCell({
-                id: 'nested-details-heading',
-                type: 'heading',
-                value: 'Registry Information',
-              })
-              .addCell({
-                type: 'subheading',
-                value: 'Registry-provided data',
-              })
-              .build()
-              .flat(1),
+            id: 'nested-details-heading',
+            type: 'heading',
+            value: 'Registry Information',
           })
           .addCell({
-            type: 'details',
-            hideSeparator: true,
-            value: {
-              data: Object.entries(pluginsOutput?.businessInformation?.data?.[0])?.map(
-                ([title, value]) => ({
-                  title,
-                  value,
-                }),
-              ),
+            id: 'nested-details-subheading',
+            type: 'subheading',
+            value: 'Registry-Provided Data',
+            props: {
+              className: 'mb-4',
             },
-            workflowId: workflow?.id,
-            documents: workflow?.context?.documents,
           })
+          .addCell(cell)
           .build()
           .flat(1),
       })
       .build();
-  }, [pluginsOutput, workflow]);
+  }, [getCell]);
 };