+++ packages/ui/src/components/organisms/DynamicForm/components/layouts/FieldLayout/FieldLayout.tsx
@@ -24,7 +24,7 @@ export const FieldLayout = ({
   const errors = useMemo(() => Array.from(new Set(rawErrors)), [rawErrors]);
 
   return (
-    <div className="flex flex-col gap-1 py-3">
+    <div className="flex flex-col gap-2 py-3">
       {isLabelEnabled ? (
         <Label htmlFor={id}>
           {label}