+++ packages/common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/common
 
+## 0.7.41
+
+### Patch Changes
+
+- updating versions
+
+## 0.7.40
+
+### Patch Changes
+
+- version bump
+
 ## 0.7.39
 
 ### Patch Changes