+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ballerine/workflow-core
 
+## 0.5.41
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.41
+
+## 0.5.40
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.7.40
+
 ## 0.5.39
 
 ### Patch Changes