+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,20 @@
 # web-ui-sdk
 
+## 1.4.38
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.41
+
+## 1.4.37
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.7.40
+
 ## 1.4.36
 
 ### Patch Changes